Reduce txn fetchSize during resend
Tessera uses a hardcoded fetch size of 10000 when retrieving
transactions during a resend request from a remote node. For databases
that contain large payload sizes in the transactions, this results in
very large repsonse times to retrieve the batch and memory exhaustion.
Memory exhaustion causes a restart of the database connection and the
offset is reset to 0, so the node starts streaming payloads from the
very beginning, which results in 500 errors on the recipient. This
commit reduces the fetch size to 1000 (with the assumption that this
will be tunable at some point in the future in tessera)